Move PagesToFileCollection to utils as a function
